fix type

+++ payment/internal/repository/repository_payment_method.go
@@ -4,7 +4,7 @@ import (
 	"github.com/GrabItYourself/giys-backend/lib/postgres/models"
 )
 
-func (r *Repository) GetPaymentMethodById(id int64) (*models.PaymentMethod, error) {
+func (r *Repository) GetPaymentMethodById(id int32) (*models.PaymentMethod, error) {
 	var paymentMethod models.PaymentMethod
 	err := r.pg.Where("id = ?", id).Take(&paymentMethod).Error
 	if err != nil {
